+++ grafana-plugin/src/pages/integrations/Integrations.tsx
@@ -51,18 +51,23 @@ class Integrations extends React.Component<IntegrationsProps, IntegrationsState>
     errorData: initErrorDataState(),
   };
 
+  private isMounted: boolean;
   private alertReceiveChanneltoPoll: { [key: string]: number } = {};
   private alertReceiveChannelTimerId: ReturnType<typeof setTimeout>;
 
   async componentDidMount() {
+    this.isMounted = false;
     this.update().then(() => this.parseQueryParams(true));
+  }
+
+  componentWillUnmount() {
+    this.isMounted = false;
+    clearInterval(this.alertReceiveChannelTimerId);
   }
 
   componentDidUpdate(prevProps: Readonly<IntegrationsProps>): void {
     if (prevProps.match.params.id && !this.props.match.params.id) {
-      this.setState({ errorData: initErrorDataState() }, () => {
-        this.parseQueryParams();
-      });
+      this.parseQueryParams();
     }
   }
 
@@ -70,7 +75,7 @@ class Integrations extends React.Component<IntegrationsProps, IntegrationsState>
     const { store, history } = this.props;
     store.selectedAlertReceiveChannel = alertReceiveChannelId;
 
-    if (shouldRedirect) {
+    if (shouldRedirect && this.isMounted) {
       history.push(`${PLUGIN_ROOT}/integrations/${alertReceiveChannelId || ''}`);
     }
   };
@@ -122,10 +127,6 @@ class Integrations extends React.Component<IntegrationsProps, IntegrationsState>
     const { store } = this.props;
     return store.alertReceiveChannelStore.updateItems();
   };
-
-  componentWillUnmount() {
-    clearInterval(this.alertReceiveChannelTimerId);
-  }
 
   render() {
     const {