+++ engine/settings/base.py
@@ -215,6 +215,7 @@ INSTALLED_APPS = [
     "apps.auth_token",
     "apps.public_api",
     "apps.grafana_plugin",
+    "apps.webhooks",
     "corsheaders",
     "debug_toolbar",
     "social_django",