+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable web overrides for Terraform-based schedules
 - Direct user paging improvements ([1358](https://github.com/grafana/oncall/issues/1358))
+- Add filtering by escalation chain to alert groups page ([1535](https://github.com/grafana/oncall/pull/1535))
 
 ## v1.1.36 (2023-03-09)